Refactor StatementServiceImpl and introduce CommandExecutionHelper for improved SQL execution handling
- Removed unused imports and fields in StatementServiceImpl.
- Introduced CommandExecutionHelper to centralize session validation, activity updates, and error handling for SQL execution.
- Updated ExecuteQueryAction and ExecuteUpdateAction to utilize the new CommandExecutionHelper for executing SQL statements with resilience.
- Enhanced ActionContext to include SqlEnhancerEngine for SQL optimization.
- Updated tests to reflect changes in the action context and command execution logic.
