+++ ojp-server/src/main/java/org/openjproxy/grpc/server/action/transaction/ExecuteQueryAction.java
@@ -1,0 +1,128 @@
+package org.openjproxy.grpc.server.action.transaction;
+
+import com.openjproxy.grpc.OpResult;
+import com.openjproxy.grpc.StatementRequest;
+import io.grpc.stub.StreamObserver;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
+import org.openjproxy.grpc.ProtoConverter;
+import org.openjproxy.grpc.dto.Parameter;
+import org.openjproxy.grpc.server.ConnectionSessionDTO;
+import org.openjproxy.grpc.server.action.Action;
+import org.openjproxy.grpc.server.action.ActionContext;
+import org.openjproxy.grpc.server.statement.StatementFactory;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.List;
+
+import static org.openjproxy.grpc.server.action.session.ResultSetHelper.handleResultSet;
+import static org.openjproxy.grpc.server.action.streaming.SessionConnectionHelper.sessionConnection;
+import static org.openjproxy.grpc.server.action.transaction.CommandExecutionHelper.executeWithResilience;
+
+@Slf4j
+public class ExecuteQueryAction implements Action<StatementRequest, OpResult> {
+
+    private static final ExecuteQueryAction INSTANCE = new ExecuteQueryAction();
+
+    /**
+     * Private constructor for singleton.
+     */
+    private ExecuteQueryAction() {
+    }
+
+    /**
+     * Returns the singleton instance of this action.
+     *
+     * @return the singleton instance
+     */
+    public static ExecuteQueryAction getInstance() {
+        return INSTANCE;
+    }
+
+    @Override
+    public void execute(ActionContext context, StatementRequest request, StreamObserver<OpResult> responseObserver) {
+        log.info("Executing query for {}", request.getSql());
+
+        executeWithResilience(context, request, responseObserver, () ->
+                        executeQueryInternal(context, request, responseObserver),
+                null, "query");
+    }
+
+    /**
+     * Internal method for executing queries without segregation logic.
+     */
+    private void executeQueryInternal(ActionContext actionContext, StatementRequest request, StreamObserver<OpResult> responseObserver)
+            throws SQLException {
+
+        ConnectionSessionDTO dto = sessionConnection(actionContext, request.getSession(), true);
+
+        // Phase 2: SQL Enhancement with timing
+        String sql = request.getSql();
+        long enhancementStartTime = System.currentTimeMillis();
+
+        var sqlEnhancerEngine = actionContext.getSqlEnhancerEngine();
+        var datasourceMap = actionContext.getDatasourceMap();
+        var sessionManager = actionContext.getSessionManager();
+
+        if (sqlEnhancerEngine.isEnabled()) {
+            // Ensure schema is loaded before enhancement (on-demand, only once)
+            try {
+                // Get the DataSource for this connection
+                String dsKey = dto.getSession().getConnHash();
+                DataSource dataSource = datasourceMap.get(dsKey);
+
+                if (dataSource != null) {
+                    // Get catalog and schema from the connection
+                    Connection connection = dto.getConnection();
+                    String catalogName = connection.getCatalog();
+                    String schemaName = connection.getSchema();
+
+                    // PostgreSQL: Use "public" schema if schema name is null or empty
+                    // This ensures tables created in the default schema are visible to Calcite
+                    if ((schemaName == null || schemaName.isEmpty()) &&
+                            connection.getMetaData().getDatabaseProductName().equalsIgnoreCase("PostgreSQL")) {
+                        schemaName = "public";
+                        log.debug("Using default PostgreSQL 'public' schema for schema loading");
+                    }
+
+                    // Ensure schema is loaded (thread-safe, idempotent)
+                    sqlEnhancerEngine.ensureSchemaLoaded(dataSource, catalogName, schemaName);
+                } else {
+                    log.debug("No DataSource found for connection hash: {}", dsKey);
+                }
+            } catch (Exception e) {
+                // Log but don't fail - enhancement can proceed without schema
+                log.warn("Failed to ensure schema loaded: {}", e.getMessage());
+            }
+
+            org.openjproxy.grpc.server.sql.SqlEnhancementResult result = sqlEnhancerEngine.enhance(sql);
+            sql = result.getEnhancedSql();
+
+            long enhancementDuration = System.currentTimeMillis() - enhancementStartTime;
+
+            if (result.isModified()) {
+                log.debug("SQL was enhanced in {}ms: {} -> {}", enhancementDuration,
+                        request.getSql().substring(0, Math.min(request.getSql().length(), 50)),
+                        sql.substring(0, Math.min(sql.length(), 50)));
+            } else if (enhancementDuration > 10) {
+                log.debug("SQL enhancement took {}ms (no modifications)", enhancementDuration);
+            }
+        }
+
+        List<Parameter> params = ProtoConverter.fromProtoList(request.getParametersList());
+        if (CollectionUtils.isNotEmpty(params)) {
+            PreparedStatement ps = StatementFactory.createPreparedStatement(sessionManager, dto, sql, params, request);
+            String resultSetUUID = sessionManager.registerResultSet(dto.getSession(), ps.executeQuery());
+            handleResultSet(actionContext, dto.getSession(), resultSetUUID, responseObserver);
+        } else {
+            Statement stmt = StatementFactory.createStatement(sessionManager, dto.getConnection(), request);
+            String resultSetUUID = sessionManager.registerResultSet(dto.getSession(),
+                    stmt.executeQuery(sql));
+            handleResultSet(actionContext, dto.getSession(), resultSetUUID, responseObserver);
+        }
+    }
+}
