+++ ojp-server/src/test/java/org/openjproxy/grpc/server/action/connection/HandleUnpooledXAConnectionActionTest.java
@@ -52,6 +52,7 @@ class HandleUnpooledXAConnectionActionTest {
                 sessionManager,
                 circuitBreakerRegistry,
                 serverConfiguration,
+                null,
                 null);
 
         ConnectionDetails connectionDetails = ConnectionDetails.newBuilder()
